requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Settings & validation
pydantic==2.9.2
pydantic-settings==2.6.1

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Auth
python-jose[cryptography]==3.3.0
httpx==0.28.0

# PDF processing — digital
pdfplumber==0.11.4

# PDF processing — scanned
pytesseract==0.3.13
Pillow==11.0.0
pdf2image==1.17.0

# PDF export
weasyprint==62.3

# Object storage
aioboto3==13.2.0

# Multipart uploads
python-multipart==0.0.12
